+++ test/deps.ts
@@ -1,1 +1,1 @@
-export * from "https://deno.land/std@0.126.0/testing/asserts.ts";
+export * from "https://deno.land/std@0.127.0/testing/asserts.ts";
